Add a project based layout
Presents projects in a flex list that utilizes two-columns for wide
displays and a single column for small ones. Each project page has a
coherent layout similar to that of an article but focussed on
presentation instead of text writing, i.e. it is wider compared to the
standard layout from the theme.