Underlay the header with the fractal image
Tried to stay somewhat accurate to the tone of the color to not hurt
readability and because I think the slate style is actually not the
worst overall. I realize it is a little gimmicky but the fractal
generation has been the frontface of `image` for a while now so let's
embrace it.