Add simple content for each project.
Simplistic but filled project pages, short summaries and title cards
should at least get the point across that the organization is not an
effort for a small number of crates but should span multiple, stronger
together than alone.

And encourage small scale support, we could get a most-recent
contributor or so.